Keep monitoring filters responsive during chart updates
